+++ src/bitwarden_exporter/__init__.py
@@ -5,9 +5,9 @@ Classes:
     BitwardenException: Base exception for Bitwarden Export.
 """
 
-from typing import Optional
+import tempfile
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 
 class BitwardenExportSettings(BaseModel):
@@ -15,23 +15,34 @@ class BitwardenExportSettings(BaseModel):
     Configuration for the Bitwarden Exporter CLI.
 
     Attributes:
-        export_location: Absolute or relative path to the output KeePass (.kdbx) file.
-        export_password: KeePass database password as plain text (read from a file if a path is supplied).
-        allow_duplicates: If True, items that belong to multiple collections will be duplicated across them.
         tmp_dir: Directory used to store temporary, sensitive artifacts (attachments, SSH keys) during export.
         debug: Enables verbose logging and keeps the temporary directory after export for troubleshooting.
         bw_executable: Path or command name of the Bitwarden CLI executable (defaults to "bw").
     """
 
-    export_location: str
-    export_password: str
-    allow_duplicates: bool
-    tmp_dir: str
-    debug: bool
+    tmp_dir: str = Field(default_factory=tempfile.mkdtemp)
+    debug: bool = False
     bw_executable: str = "bw"
 
 
-class BitwardenException(Exception):
-    """
-    Base Exception for Bitwarden Export
-    """
+BITWARDEN_EXPORTER_GLOBAL_SETTINGS: BitwardenExportSettings = BitwardenExportSettings()
+
+APPLICATION_NAME_ASCII = r"""
+ ____  _ _                         _
+| __ )(_) |___      ____ _ _ __ __| | ___ _ __
+|  _ \| | __\ \ /\ / / _` | '__/ _` |/ _ \ '_ \
+| |_) | | |_ \ V  V / (_| | | | (_| |  __/ | | |
+|____/|_|\__| \_/\_/ \__,_|_|_ \__,_|\___|_| |_|
+| ____|_  ___ __   ___  _ __| |_ ___ _ __
+|  _| \ \/ / '_ \ / _ \| '__| __/ _ \ '__|
+| |___ >  <| |_) | (_) | |  | ||  __/ |
+|_____/_/\_\ .__/ \___/|_|   \__\___|_|
+           |_|
+"""
+
+CLI_DEBUG_HELP = """
+Enable verbose logging, This will print debug logs, THAT MAY CONTAIN SENSITIVE INFORMATION,
+This will not delete the temporary directory after the export.
+"""
+
+APPLICATION_PACKAGE_NAME = "bitwarden-exporter"