+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.10.0](https://github.com/arpanrec/bitwarden-exporter/compare/1.9.1...1.10.0) (2025-10-29)
+
+
+### Features
+
+* add typer framework ([#30](https://github.com/arpanrec/bitwarden-exporter/issues/30)) ([6d8e251](https://github.com/arpanrec/bitwarden-exporter/commit/6d8e25192664169a204fa85a40babd6f22e71322))
+
+## [1.9.1](https://github.com/arpanrec/bitwarden-exporter/compare/1.9.0...1.9.1) (2025-10-27)
+
+
+### Bug Fixes
+
+* add shorthand options for CLI arguments in `keepass` command ([bb0c7e0](https://github.com/arpanrec/bitwarden-exporter/commit/bb0c7e0f2f127b47d278547d46f8d31e8e412a61))
+
 # [1.9.0](https://github.com/arpanrec/bitwarden-exporter/compare/1.8.1...1.9.0) (2025-10-27)